+++ public/data/commentData.json
@@ -13,7 +13,7 @@
   },
   {
     "id": 3,
-    "userName": "Beenzino",
+    "userName": "B",
     "content": "ily",
     "isLiked": true
   }
