+++ src/components/CommentBoard.js
@@ -1,20 +1,32 @@
 import React from 'react';
-import CommentItem from './CommentItem';
+import '../pages/bohyun/Main/main.scss';
 
-const CommentBoard = ({ commentList }) => {
-  return (
-    <div>
-      <CommentItem />
-      {commentList.map(comment => {
-        // return <CommentItem comment={comment} />;
-        return (
-          <div className=".commentItem">
-            <h1>{comment}</h1>
+const CommentBoard = ({ commentList, changeHeart, heart, setCommentList }) => {
+  return commentList.map((comment, index) => {
+    return (
+      <div className="newCommentBox">
+        <div className="newCommentEntry" key={index}>
+          <div className="entry">
+            <span className="boldID"> arielle</span>
+            <span>{comment}</span>
           </div>
-        );
-      })}
-    </div>
-  );
+          <div className="newCommentButtons">
+            <img onClick={changeHeart} src={heart} className="entryHeart" />
+            <button
+              className="deleteButton"
+              onClick={() => {
+                let copy = [...commentList];
+                copy.splice(index, 1);
+                setCommentList(copy);
+              }}
+            >
+              x
+            </button>
+          </div>
+        </div>
+      </div>
+    );
+  });
 };
 
 export default CommentBoard;
