+++ src/pages/bohyun/Main/Main.js
@@ -1,19 +1,17 @@
-import React, { useState, useEffect } from 'react';
-import CommentBoard from '../../../components/CommentBoard';
-import CommentBox from '../../../components/CommentBox';
+import React, { useState } from 'react';
+import Feeds from '../../../components/Feeds';
 import './main.scss';
 
 const MainBohyun = () => {
   const [inputValue, setInputValue] = useState('');
   const [commentList, setCommentList] = useState([]);
+
   const addComment = e => {
     e.preventDefault();
     setCommentList([...commentList, inputValue]);
     setInputValue('');
   };
-  // useEffect(() => {
-  //   commentList;
-  // }, [commentList]);
+
   const [heart, setHeart] = useState(
     require('../../../assets/bohyun/icons/heart.png')
   );
@@ -41,23 +39,25 @@ const MainBohyun = () => {
         </div>
 
         <div className="navigate">
-          <a href="">
-            {' '}
+          <a>
             <img
               src={require('../../../assets/bohyun/icons/explore.png')}
               className="icons"
+              alt="exloreIcon"
             />
           </a>
-          <a href="">
+          <a>
             <img
               src={require('../../../assets/bohyun/icons/heart.png')}
               className="icons"
+              alt="heartIcon"
             />
           </a>
-          <a href="">
+          <a>
             <img
               src={require('../../../assets/bohyun/icons/profile.png')}
               className="icons"
+              alt="profileIcon"
             />
           </a>
         </div>
@@ -65,120 +65,14 @@ const MainBohyun = () => {
 
       <main className="mainbody">
         <div className="feeds">
-          <div className="article">
-            <div className="user_whole">
-              <img
-                src={require('../../../assets/bohyun/img_profile/profile_left_top.JPG')}
-                width="20px"
-                className="circle_maintop"
-              />
-              <div className="user_info">
-                <p>arielle</p>
-              </div>
-              <div className="dropdown">
-                <i className="fa-solid fa-ellipsis" />
-              </div>
-            </div>
-            <img
-              src={require('../../../assets/bohyun/img_profile/main.JPG')}
-              className="main_img"
-            />
-            <div className="buttons_whole">
-              <div className="buttons_left">
-                <i className="fa-solid fa-heart" />
-                <i className="fa-solid fa-comment" />
-                <i className="fa-solid fa-paper-plane" />
-              </div>
-              <div className="buttons_right">
-                <i className="fa-solid fa-bookmark" />
-              </div>
-            </div>
-            <div className="likes">
-              <img
-                src={require('../../../assets/bohyun/img_profile/profile_left_bottom.jpeg')}
-                className="circle_mainbottom"
-              />
-              <div className="like_counts">
-                <p>
-                  Liked by <span className="boldID">realisshoman</span> and
-                  others
-                </p>
-              </div>
-            </div>
-            <div className="descriptions">
-              <span className="boldID">arielle</span>
-
-              <div className="caption">
-                <p>
-                  So this is it ! 🙌🏽 I’ve just landed from Milan and that wraps
-                  up my flying life as Emirates cabin crew 👠 I have been super
-                  blessed and lucky enough to travel to 41 countries and 71
-                  cities around the world meeting amazing people along the way
-                  🌏
-                </p>
-              </div>
-            </div>
-            <div className="friendComment_whole">
-              <div className="friendComment_user">
-                <span className="boldID">realisshoman</span>
-                <div className="friendComment_redHeart">
-                  <p>❤️❤️</p>
-                </div>
-              </div>
-              <div className="friendComment_heart">
-                <img
-                  src={require('../../../assets/bohyun/icons/heart.png')}
-                  width="13px"
-                />
-              </div>
-            </div>
-            <CommentBox />
-            <div className="newCommentBox">
-              {commentList.map((event, index) => {
-                // setCommentList(event.target.value);
-                return (
-                  <div className="newCommentEntry" key={index}>
-                    <div className="entry">
-                      <span className="boldID"> arielle</span>
-                      <span>{event}</span>
-                    </div>
-                    <div className="newCommentButtons">
-                      <img
-                        onClick={changeHeart}
-                        src={heart}
-                        className="entryHeart"
-                      />
-                      <button
-                        className="deleteButton"
-                        onClick={() => {
-                          let copy = [...commentList];
-                          copy.splice(index, 1);
-                          setCommentList(copy);
-                        }}
-                      >
-                        x
-                      </button>
-                    </div>
-                  </div>
-                );
-              })}
-            </div>
-
-            <form onSubmit={addComment} className="commentsWrapper">
-              <div className="comments_box">
-                <input
-                  value={inputValue}
-                  type="text"
-                  placeholder="comment"
-                  className="comments-box"
-                  onChange={event => setInputValue(event.target.value)}
-                />
-              </div>
-              <div className="comments_button">
-                <button id="postButton">post</button>
-              </div>
-            </form>
-          </div>
+          <Feeds
+            commentList={commentList}
+            changeHeart={changeHeart}
+            heart={heart}
+            setCommentList={setCommentList}
+            addComment={addComment}
+            inputValue={inputValue}
+          />
         </div>
 
         <div className="main_right">
@@ -186,6 +80,7 @@ const MainBohyun = () => {
             <img
               src={require('../../../assets/bohyun/img_profile/beenzino.JPG')}
               className="right_profile_pic"
+              alt="Profile shown on the top right"
             />
 
             <div className="right_profile_text">
@@ -208,6 +103,7 @@ const MainBohyun = () => {
                   <img
                     src={require('../../../assets/bohyun/img_stories/story1.jpeg')}
                     className="round_img"
+                    alt="1st friend's story"
                   />
                 </div>
                 <div className="right_stories_text">
@@ -222,6 +118,7 @@ const MainBohyun = () => {
                   <img
                     src={require('../../../assets/bohyun/img_stories/story2.jpeg')}
                     className="round_img"
+                    alt="2nd friend's story"
                   />
                 </div>
                 <div className="right_stories_text">
@@ -236,6 +133,7 @@ const MainBohyun = () => {
                   <img
                     src={require('../../../assets/bohyun/img_stories/story3.JPG')}
                     className="round_img"
+                    alt="3th friend's story"
                   />
                 </div>
                 <div className="right_stories_text">
@@ -250,6 +148,7 @@ const MainBohyun = () => {
                   <img
                     src={require('../../../assets/bohyun/img_stories/story4.jpeg')}
                     className="round_img"
+                    alt="4th friend's story"
                   />
                 </div>
                 <div className="right_stories_text">
@@ -275,6 +174,7 @@ const MainBohyun = () => {
                 <img
                   src={require('../../../assets/bohyun/img_recommendations/rec1.JPG')}
                   className="round_img"
+                  alt="recommended user1"
                 />
                 <div className="right_recommendations_text">
                   <p>nasa</p>
@@ -289,6 +189,7 @@ const MainBohyun = () => {
                 <img
                   src={require('../../../assets/bohyun/img_recommendations/rec2.JPG')}
                   className="round_img"
+                  alt="recommended user2"
                 />
                 <div className="right_recommendations_text">
                   <p>gigihadid</p>
@@ -303,6 +204,7 @@ const MainBohyun = () => {
                 <img
                   src={require('../../../assets/bohyun/img_recommendations/rec3.jpeg')}
                   className="round_img"
+                  alt="recommended user3"
                 />
                 <div className="right_recommendations_text">
                   <p>tatemodern</p>
